Refactor js and templatetags of pages
This allows the scripts for drag & drop and the template filters to be re-used for other tree models (e.g. the language tree node)
