+++ src/cms/templates/pages/page_form.html
@@ -17,7 +17,9 @@
 {% block javascript_head %}
     <script src="{% static 'js/confirmation-popups.js' %}"></script>
     <script src="{% static 'js/custom.js' %}"></script>
-    <script src="{% static 'js/page.js' %}"></script>
+    <script src="{% static 'js/pages/page_mirroring.js' %}"></script>
+    <script src="{% static 'js/pages/page_permissions.js' %}"></script>
+    <script src="{% static 'js/pages/page_side_by_side.js' %}"></script>
     <script src="{% static 'tinymce/tinymce.min.js' %}"></script>
     <script src="{% static 'tinymce/themes/silver/theme.js' %}"></script>
     <script src="{% static 'tinymce/plugins/paste/plugin.js' %}"></script>
@@ -425,5 +427,3 @@ document.addEventListener('DOMContentLoaded', function(){
 }, false);
 </script>
 {% endblock %}
-
-
