+++ src/cms/templatetags/page_filters.py
@@ -18,17 +18,3 @@ def get_last_root_page(pages):
     :rtype: ~cms.models.pages.page.Page
     """
     return pages.filter(parent=None).last()
-
-
-@register.filter
-def get_descendants(page):
-    """
-    This filter returns the ids of all the page's descendants.
-
-    :param page: The requested page
-    :type page: ~cms.models.pages.page.Page
-
-    :return: The list of all the page's descendants' ids
-    :rtype: list [ int ]
-    """
-    return [descendant.id for descendant in page.get_descendants(include_self=True)]
